// src/app/core/api/generated/model/category-response.ts
/**
 * Seggi App API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CategoryResponse { 
    idCategory?: number;
    name?: string;
    description?: string;
}